feat: add run_peer_review and get_peer_reviews MCP tools (#770)
The peer review generation logic was inlined in the API route handler,
making it inaccessible to MCP clients. Extract it into a shared service
and expose two new MCP tools.

Changes:
- Create src/lib/ai/peer-review-service.ts with runPeerReview() function
- Simplify POST /peer-review route to delegate to the service
- Add run_peer_review MCP tool (triggers full peer review)
- Add get_peer_reviews MCP tool (reads stored reviews)

Fixes #770
